refactor: extract caching logic - intermediate step before Planner-Executor pattern
- Move content processing methods to ContentProcessor
- Extract cache orchestration from GeminiClient
- Remove deprecated methods related to caching and content part identification
- Prepare for transition to data-driven cache planning
