+++ gemini_batch/client/cache_manager.py
@@ -11,10 +11,13 @@ from datetime import datetime, timedelta, timezone
 import hashlib
 import logging
 import time
-from typing import Dict, List, Optional, Union
+from typing import Callable, Dict, List, Optional, Union
 
 from google import genai
 from google.genai import types
+
+from gemini_batch.client.content_processor import ContentProcessor
+from gemini_batch.exceptions import APIError
 
 from ..config import ConfigManager
 
@@ -80,11 +83,13 @@ class CacheManager:
         config_manager: ConfigManager,
         token_counter,
         default_ttl_seconds: int = 3600,
+        content_processor: ContentProcessor = None,
     ):
         self.client = client
         self.config_manager = config_manager
         self.token_counter = token_counter
         self.default_ttl_seconds = default_ttl_seconds
+        self.content_processor = content_processor
 
         # Cache tracking
         self.active_caches: Dict[str, CacheInfo] = {}
@@ -180,6 +185,79 @@ class CacheManager:
                 fallback_required=True,
                 error=f"Cache operation failed: {e}",
             )
+
+    def generate_with_cache(
+        self,
+        parts: List[types.Part],
+        raw_generation_func: Callable[[], types.GenerateContentResponse],
+        system_instruction: Optional[str] = None,
+        **options,
+    ) -> types.GenerateContentResponse:
+        """
+        Orchestrates a cache-aware generation call.
+        Returns a raw GenerateContentResponse for the client to process.
+        """
+        # 1. Analyze content and determine the appropriate caching strategy.
+        token_count = self.token_counter.count_tokens(self.config_manager.model, parts)
+        cache_analysis = self.config_manager.can_use_caching(self.config_manager.model, token_count)
+
+        if not cache_analysis.get("supported"):
+            return raw_generation_func()
+
+        strategy = cache_analysis["recommendation"]
+        log.debug("Cache strategy determined: %s", strategy)
+
+        # 2. Delegate to the appropriate handler based on the strategy.
+        try:
+            if strategy == "explicit":
+                # This helper now relies on ContentProcessor and returns a raw response.
+                return self._handle_explicit_cache(parts, system_instruction, **options)
+
+            # Note: Implicit caching is automatic. The raw_generation_func will handle it
+            # if the parts are structured correctly. We just need to prepare them.
+            elif strategy == "implicit":
+                optimized_parts = self.content_processor.optimize_for_implicit_cache(parts)
+                # The client's raw_generation_func will use these optimized parts.
+                return raw_generation_func(parts_override=optimized_parts)
+
+        except APIError as e:
+            log.warning("Cache generation failed, falling back. Reason: %s", e)
+            return raw_generation_func()
+
+        # 3. Fallback for any strategy that isn't explicitly handled (e.g., "none").
+        return raw_generation_func()
+
+    def _handle_explicit_cache(
+        self,
+        parts: List[types.Part],
+        system_instruction: Optional[str],
+        **options,
+    ) -> types.GenerateContentResponse:
+        """Handles explicit cache generation and returns the raw API response."""
+
+        # A. Get or create the cache object.
+        cached_content_obj = self.create_or_get(
+            parts, self.config_manager.model, strategy="explicit"
+        )
+        log.info("Using explicit cache: %s", cached_content_obj.name)
+
+        # B. Delegate content separation to ContentProcessor.
+        _, prompt_parts = self.content_processor.separate_cacheable_content(parts)
+
+        # C. Make the cache-specific API call.
+        model = self.client.get_generative_model(model_name=self.config_manager.model)
+        cached_content = model.from_cached_content(cached_content_name=cached_content_obj.name)
+
+        # The CacheManager is responsible for THIS specific API call.
+        # It does NOT need a generic retry mechanism, as failure here is a cache failure, not a generic API failure.
+        response = cached_content.generate_content(prompt_parts, **options)
+
+        # D. Return the raw response. The GeminiClient is responsible for processing it.
+        # We can enhance it with our cache info.
+        if hasattr(response, "usage_metadata"):
+             response.usage_metadata.cached_content_token_count = cached_content_obj.token_count
+
+        return response
 
     def _get_existing_cache(self, content_hash: str) -> Optional[CacheResult]:
         """Check for existing valid cache"""
